REFAC:  Moved some functions to the review queue loader
